UCS/SIGNAL: signal and sigaction handlers were moved to separate module
- added separate module to handle signal and sigaction calls
