+++ src/ucs/configure.m4
@@ -133,20 +133,6 @@ AS_IF([test "x$with_bfd" != xno],
 
 
 #
-# Disable overriding sigaction
-#
-AC_ARG_ENABLE([sigaction-override],
-    AS_HELP_STRING([--disable-sigaction-override],
-                   [Disable sigaction,signal function overriding, default: NO]),
-    [],
-    [enable_sigaction_override=yes])
-
-AS_IF([test "x$enable_sigaction_override" = xyes],
-      AC_DEFINE([ENABLE_SIGACTION_OVERRIDE], [1], [Enable sigaction,signal function overriding])
-    )
-
-
-#
 # Enable statistics and counters
 #
 AC_ARG_ENABLE([stats],
@@ -302,4 +288,5 @@ AC_CHECK_FUNCS([__clear_cache], [], [])
 AC_CHECK_FUNCS([__aarch64_sync_cache_range], [], [])
 
 
-AC_CONFIG_FILES([src/ucs/Makefile])
+AC_CONFIG_FILES([src/ucs/Makefile
+                 src/ucs/signal/Makefile])
