+++ src/ucs/debug/debug.c
@@ -192,8 +192,6 @@ extern char *cplus_demangle(const char *, int);
 #endif
 
 static int ucs_debug_backtrace_is_excluded(void *address, const char *symbol);
-static int orig_sigaction(int signum, const struct sigaction *act,
-                          struct sigaction *oact);
 
 
 static char *ucs_debug_strdup(const char *str)
@@ -902,7 +900,7 @@ static void ucs_error_freeze(const char *message)
 
     /* restore original SIGINT handler to allow termitate process via Ctrl+C */
     sigemptyset(&sigact.sa_mask);
-    orig_sigaction(SIGINT, &sigact, NULL);
+    sigaction(SIGINT, &sigact, NULL);
 
     ucs_debug_stop_other_threads();
 
@@ -1096,32 +1094,12 @@ int ucs_debug_is_handle_errors()
     return ucs_global_opts.handle_errors & mask;
 }
 
-static void* ucs_debug_get_orig_func(const char *symbol, void *replacement)
-{
-    void *func_ptr;
-
-    func_ptr = dlsym(RTLD_NEXT, symbol);
-    if (func_ptr == NULL) {
-        func_ptr = dlsym(RTLD_DEFAULT, symbol);
-    }
-    return func_ptr;
-}
-
-#ifdef ENABLE_SIGACTION_OVERRIDE
-#if !HAVE_SIGHANDLER_T
-#if HAVE___SIGHANDLER_T
-typedef __sighandler_t *sighandler_t;
-#else
-#error "Port me"
-#endif
-#endif
-
-static int ucs_debug_is_error_signal(int signum)
+int ucs_debug_is_error_signal(int signum)
 {
     khiter_t hash_it;
     int result;
 
-    if (!ucs_debug_is_handle_errors()) {
+    if (!ucs_debug_initialized || !ucs_debug_is_handle_errors()) {
         return 0;
     }
 
@@ -1131,48 +1109,6 @@ static int ucs_debug_is_error_signal(int signum)
     result = (hash_it != kh_end(&ucs_signal_orig_action_map));
     ucs_recursive_spin_unlock(&ucs_kh_lock);
     return result;
-}
-
-sighandler_t signal(int signum, sighandler_t handler)
-{
-    typedef sighandler_t (*sighandler_func_t)(int, sighandler_t);
-
-    static sighandler_func_t orig = NULL;
-
-    if (ucs_debug_initialized && ucs_debug_is_error_signal(signum)) {
-        return SIG_DFL;
-    }
-
-    if (orig == NULL) {
-        orig = (sighandler_func_t)ucs_debug_get_orig_func("signal", signal);
-    }
-
-    return orig(signum, handler);
-}
-
-int sigaction(int signum, const struct sigaction *act, struct sigaction *oact)
-{
-    if (ucs_debug_initialized && ucs_debug_is_error_signal(signum)) {
-        return orig_sigaction(signum, NULL, oact); /* Return old, do not set new */
-    }
-
-    return orig_sigaction(signum, act, oact);
-}
-
-#endif /* ENABLE_SIGACTION_OVERRIDE */
-
-static int orig_sigaction(int signum, const struct sigaction *act,
-                          struct sigaction *oact)
-{
-    typedef int (*sigaction_func_t)(int, const struct sigaction*, struct sigaction*);
-
-    static sigaction_func_t orig = NULL;
-
-    if (orig == NULL) {
-        orig = (sigaction_func_t)ucs_debug_get_orig_func("sigaction", sigaction);
-    }
-
-    return orig(signum, act, oact);
 }
 
 static void ucs_debug_signal_handler(int signo)
@@ -1252,9 +1188,13 @@ static void ucs_set_signal_handler(void (*handler)(int, siginfo_t*, void *))
     }
     sigemptyset(&sigact.sa_mask);
 
+    /* This function must be called prior to debug infra is initialized
+     * to forbid processing signal by internal signal handler */
+    ucs_assert(!ucs_debug_initialized);
+
     for (i = 0; i < ucs_global_opts.error_signals.count; ++i) {
-        ret = orig_sigaction(ucs_global_opts.error_signals.signals[i], &sigact,
-                             &old_action);
+        ret = sigaction(ucs_global_opts.error_signals.signals[i], &sigact,
+                        &old_action);
         if (ret < 0) {
             ucs_warn("failed to set signal handler for sig %d : %m",
                      ucs_global_opts.error_signals.signals[i]);
@@ -1305,7 +1245,7 @@ void ucs_debug_init()
         memset(&sigact, 0, sizeof(sigact));
         memset(&old_action, 0, sizeof(old_action));
         sigact.sa_handler = ucs_debug_signal_handler;
-        orig_sigaction(ucs_global_opts.debug_signo, &sigact, &old_action);
+        sigaction(ucs_global_opts.debug_signo, &sigact, &old_action);
         ucs_debug_save_original_sighandler(ucs_global_opts.debug_signo, &old_action);
     }
 
@@ -1352,12 +1292,14 @@ static inline void ucs_debug_disable_signal_nolock(int signum)
     }
 
     original_action = kh_val(&ucs_signal_orig_action_map, hash_it);
-    ret = orig_sigaction(signum, original_action, &ucs_action);
+    /* Remove signal from map prior to call sigaction to allow process it by
+     * original sigaction */
+    kh_del(ucs_signal_orig_action, &ucs_signal_orig_action_map, hash_it);
+    ret = sigaction(signum, original_action, &ucs_action);
     if (ret < 0) {
         ucs_warn("failed to set signal handler for sig %d : %m", signum);
     }
 
-    kh_del(ucs_signal_orig_action, &ucs_signal_orig_action_map, hash_it);
     ucs_free(original_action);
 }
