+++ src/ucs/debug/debug_int.h
@@ -127,4 +127,13 @@ void ucs_handle_error(const char *message);
 const char *ucs_debug_get_symbol_name(void *address);
 
 
+/**
+ * Check if signal should be processed by UCX as error.
+ *
+ * @param signum         Signal number to check.
+ *
+ * @return 1 if signal should be processes by UCX, 0 if passed to system.
+ */
+int ucs_debug_is_error_signal(int signum);
+
 #endif
